Remove processByTimeBlockTag circularity
